Add consideration of bar alpha to xrdb and make many small fixes
This commit really should have been several much smaller commits, but
it's too late for doing so to be an effective use of time. Lots of
style fixes have been made and some code has been moved around, but
the only functional changes are the addition of alpha consideration
for xrdb updates and removal of clickable status bar.