Improve cosmetic appeal
The odd little box in the top left corner of a tag label which indicated
that there were windows with that tag has been replaced by a sleep and
trendy bar on the bottom.

Tags have been given names and the layout indicators have been more
mysterious (cool!). I have also made it so that the awful GTK window
which pop ups when Steam would like to update is masked to tag 6,
changed the dmenu hotkey to start+single-quote, and change the font
from Andale Mono to Ubuntu Mono